+++ tensorflow_decision_forests/keras/__init__.py
@@ -95,3 +95,4 @@ get_worker_idx_and_num_workers = core.get_worker_idx_and_num_workers
 build_default_input_model_signature = core.tf_core.build_default_input_model_signature
 build_default_feature_signature = core.tf_core.build_default_feature_signature
 set_training_logs_redirection = core.tf_logging.set_training_logs_redirection
+yggdrasil_model_to_keras_model = core.yggdrasil_model_to_keras_model